chore(cli): restore the file-size ratchet on main — move the /models parser, extract config tests
Main's ratchet went red after the #258/#259/#261 merges landed together:
command_deck.rs hit 5380 (allowlisted limit 5342) and config.rs 1526
(1500 cap).

- ModelsCommand + parse_models_command move next to run_refresh in
  model_catalog.rs (they are /models vocabulary, not deck plumbing),
  taking their test along: command_deck.rs 5380 → 5318.
- config.rs's test module is extracted to config_tests.rs via the same
  #[path] pattern #254 used for agent_tests.rs, and the openrouter-pin
  regression folds into the adjacent stale-pin test as a second
  scenario: config.rs 1526 → 1139.

No behavior change.
