+++ stella-cli/src/model_catalog.rs
@@ -911,6 +911,42 @@ async fn refresh_with_store(
     }
 }
 
+/// An argument-carrying form of the `/models` chat command — handled
+/// model-free by both chat surfaces: when the configured model itself is
+/// broken, `/models refresh` is how the user digs out, and routing it into
+/// a model turn fails on the very error being fixed. Parsed conservatively
+/// — a single recognized token (plus `refresh --force`); anything
+/// sentence-like stays a prompt, matching the "`/init do the thing` is a
+/// model prompt" rule.
+pub enum ModelsCommand {
+    /// `/models refresh [--force]` — re-sync the catalog, no model call.
+    Refresh { force: bool },
+    /// `/models list` — the same listing the bare `/models` prints.
+    List,
+    /// `/models <typo>` — one unrecognized token: a mistyped subcommand,
+    /// answered with usage instead of a wasted model call.
+    Usage(String),
+}
+
+/// Parse `trimmed` as a [`ModelsCommand`]; `None` leaves it on the normal
+/// path (custom expansion, then prompt).
+pub fn parse_models_command(trimmed: &str) -> Option<ModelsCommand> {
+    let (head, rest) = trimmed.split_once(char::is_whitespace)?;
+    let rest = rest.trim();
+    if head != "/models" || rest.is_empty() {
+        return None;
+    }
+    let mut words = rest.split_whitespace();
+    match (words.next(), words.next(), words.next()) {
+        (Some("refresh"), None, None) => Some(ModelsCommand::Refresh { force: false }),
+        (Some("refresh"), Some("--force"), None) => Some(ModelsCommand::Refresh { force: true }),
+        (Some("list"), None, None) => Some(ModelsCommand::List),
+        (Some(word), None, None) => Some(ModelsCommand::Usage(word.to_string())),
+        // A sentence after `/models` stays a prompt.
+        _ => None,
+    }
+}
+
 /// `stella models refresh [--force]`: the master list, then every
 /// configured provider's own live `/models` listing.
 pub async fn run_refresh(force: bool) -> Result<(), String> {
@@ -1089,6 +1125,35 @@ mod tests {
     use super::*;
     use std::collections::BTreeMap;
     use stella_model::modelsdev::{ModelCost, ModelEntry, ModelLimit, ProviderEntry};
+
+    #[test]
+    fn models_command_parses_arg_forms_and_leaves_sentences_as_prompts() {
+        assert!(matches!(
+            parse_models_command("/models refresh"),
+            Some(ModelsCommand::Refresh { force: false })
+        ));
+        assert!(matches!(
+            parse_models_command("/models refresh --force"),
+            Some(ModelsCommand::Refresh { force: true })
+        ));
+        assert!(matches!(
+            parse_models_command("/models list"),
+            Some(ModelsCommand::List)
+        ));
+        // One unrecognized token is a typo'd subcommand → usage, never a
+        // model call; a sentence stays a prompt.
+        assert!(matches!(
+            parse_models_command("/models refrsh"),
+            Some(ModelsCommand::Usage(_))
+        ));
+        assert!(parse_models_command("/models what can I use").is_none());
+        // Bare forms and non-command paths are not arg commands — and the
+        // removed `/model-<role>` heads don't parse (model config lives on
+        // the SETTINGS tab).
+        assert!(parse_models_command("/models").is_none());
+        assert!(parse_models_command("/model-default zai/glm-5.2").is_none());
+        assert!(parse_models_command("/src/main.rs explain").is_none());
+    }
 
     #[test]
     fn provider_ids_map_google_vertex_bedrock_and_pass_the_rest_through() {
